ldscripts: Fix PIE build of crt1.s.o
uClibc provides a variant of crt1 for position independent executables.
Normally that variant is called Scrt1.o. For consistency it is called
crt1.s.o in L4Re. The crt1.S file is the source for both the normal and
the position independent binary. The preprocessor selects the correct
variant of relevant code paths based on the presence of the L_Scrt1
define which must be set when building the PIE version.

Add the define so that the right version of the source is selected by
the preprocessor.

Change-Id: Idf4eb60ad2e31a7a081609b6846dfcf7859a96a0