l4re: rewrite kip address handling
Having a global magic __L4_KIP_ADDR__ symbol for the address of the KIP
does not bode well with relocatable executables. The symbol will be
shifted too by the load address delta and will not match the actual KIP
mapping of the loader.

To prevent any relocation, the KIP pointer is stored in the l4sys
library as compile time constant. This is done also in anticipation of
future features, such as ASLR and no-MMU systems without virtual
memory. The KIP address is provided by a new l4_kip() call. It will
fall back to __L4_KIP_ADDR__ for non-PIC code to keep the overhead low
where possible.

The KIP pointer is only defined in the static l4sys libraries to
satisfy static linking. For shared linking the l4_global_kip definition
ends up in libld-l4.so (via libl4sys-direct.a).

Also resolve the KIP address in the .rol4re_elf_aux section at compile
time. Otherwise the compiler will leave a relocation in the read-only
section which cannot be resolved at runtime for PIE binaries and that
will let the linking fail.

Change-Id: Ia3000a9ab3de16cca21412bf2d348195a86ebeaa