Add support for static PIE on arm(64) and x86_64
In case of a static position independent executable the startup code
must apply the relocations. This is very similar to ld.so and the
relevant code can indeed be reused. The relocation must be done before
any global symbol can be used.

If uclibc is compiled with __PIC__ the static library will have a
hidden _reloc_static_pie symbol that is normally called by the uclibc
startup code. If the final binary is linked *without* -pie the symbol
will be intercepted by a stub in crt1.o. For the static-pie case a
dedicated crt1.p.o variant is used that does not have _reloc_static_pie
which will then be taken from libc.a. The linkage of shared libc stays
the same. Depending on these three variants crt1 must use either
PC-relative addressing (PIE), lookup the GOT (shared libc) or can use
symbols directly (static).

Change-Id: I331cbbbe0bcd4cb89b1e5efdf5397def060a33d5