libloader: add support for relocation
Adds a 'base' and a 'ldr_base' member to the Prog_start_info structure.
The 'base' member specifies the load offset for the loaded executable
(usually l4re_kernel). The 'ldr_base' member is forwarded in the
l4re_aux_t aux vector entry and picked up by l4re_kernel to apply this
offset in the inner loader.

Both settings can be set in a ned script when starting L4Re programs:

  L4.default_loader:start(
    {
      base = 0x100000,        -- Offset for l4re_kernel
      ldr_base = 0x05000000,  -- Offset of hello
    },
    "rom/hello");

Change-Id: I0433a597992953f6dfcb221f2f093e2314eaae6b