moe: reserve stack top for argc
Moe uses the uclibc startup but sets up its own stack that is
allocated in the bss section. The startup code of uclibc expects argc
to be the current stack top. On most architectures this word will be
popped off the stack.

Because there was originally nothing on the stack, popping off argc will
move the stack pointer out of the allocated range. Worse, the next push
to the stack will overwrite the next element in the bss section.

Change-Id: I25e5aea16462488161b7cd414af5cddeae11cc21