l4re: add Dataspace::map_to()
This adds a method to map parts of a dataspace to a foreign task.
Despite that it takes an additional parameter for the task capability
it works the same as map_region().

This method fills the gap of mapping pages to a foreign task without
actually having them mapped in the calling task. This is required to
eagerly map pages in a loader of a newly created child task.

Change-Id: Ie1911b5024830d2c08c9e22d17eb1d62fc1efa91