sigma0: allow updating a region's rights during reserve()
The way sigma0 handles reserving regions which are an exact overlap of
an existing region is inconsistent.

This fix addresses points 1) and 2) below and allows updating a reserved
region's rights when the region is an exact overlap and the owner is not
set or is the same owner.

1) reserving an existing region with no owner for a new region with
different owner and different rights only updates the owner and not the
rights. This seems wrong, the rights should be updated.

2) reserving an existing region that has an owner, with the same owner
and different rights preserves the existing region. This seems unusual
since later regions usually take precedence over earlier ones but here
we have an exception.

Fixes #CD-409.

Change-Id: I928ccf978722dad4896ef8a50140c268d523d6d6