sigma0: fix reserve() out of memory handling
This change makes it such that reserve() returns an error as soon as it
detects that an error is returned from add(), including an out of memory
condition.

In case of an error an effort is made to revert the changes reserve()
makes as it processes overlapping regions. This can prevent existing
regions from being completely dropped from the tree if add() fails.
Since reserve() processes overlapping regions in several passes not all
overlap scenarios can necessarily be reverted.

Fixes ticket #CD-398.

Change-Id: Ie912a8cd18ea1492090bab1c5e7f7f3857feb6ab