sigma0: fix alloc_from() out of memory handling
This change makes it such that alloc_from() returns an error as soon as
it detects that an error is returned from add(), including an out of
memory condition.

In case of an error an effort is made to revert the changes alloc_from()
makes as it processes overlapping regions. This can prevent existing
regions from being completely dropped from the tree if add() fails.

Fixes ticket #CD-398.

Change-Id: I2b676c580ae702896449b5d5ecec057ba87f5dbf