l4vfs: extend L4Re::Vfs::Fs::set_fd by return code
The function must not perform any change if the passed file descriptor
is invalid (out of file descriptor range) but has to return an error
code instead. We now use a pair of a pointer to the file object that was
previously assigned to the file descriptor plus a return value to
propagate the error to the caller.

See ticket #CD-68.

Change-Id: I5c82653c3ea229dd2c11ec81e63a5177789d322d