l4re_kernel: respect eager mapping ldr_flag for stack
If the 'ldr_flags' of a newly started application contains the flags
L4RE_AUX_LDR_FLAG_EAGER_MAP, load the L4Re loader sections eager.

This is already done for the program sections (prog_attach_ds()) but
we shall do this also for the application stack. Implement this.

See ticket #CD-711.

Change-Id: I31276c669ad2558ef529833aeba915a9c7411108