moe: be careful if kernel does not have a mapdb
The presence of the mapdb is announced by the kernel since
I82ad1f74829e5e587734be67c6aeead5bd2e5010 in the KIP feature strings.
If Fiasco has no mapdb, we cannot be sure that there are no remaining
mappings if a dataspace is destroyed. Hence we cannot safely free
memory in the absence of the mapdb.

Requires-Change-Id: I82ad1f74829e5e587734be67c6aeead5bd2e5010
Change-Id: I785db5061e90ecf92e8b9753be858f1e39dc667e