moe: fold Loader::exec into Loader::start
The Loader::start method is just a tiny wrapper around Loader::exec,
which is otherwise unused. This commit removes the wrapping.

Also exec is declared as returning bool but then does `return 0;` and
ignores the return value of launch. And start declares an unused Dbg
object. This commit cleans up these things.

Change-Id: Ief316f42af3d3cb1993cb12f39577a06417169cc