sigma0/moe: remove ARM64 code for patching the syscall inst
On ARM64 the actual syscall instruction to be used depends on the
execution level the kernel is executed on. The kernel provides the
instruction on the KIP but Moe/Sigma0 would require an initial syscall
to map the KIP, see Idc858df300a9eebb6f1264b4b0957b17bdbd3f49.

The config option CONFIG_KERNEL_CPU_VIRT controls which instruction
is compiled into the binary. Runtime patching will not work due to
the W^X feature for userland.

Hence remove the code for patching the syscall instruction.

Change-Id: I86030f55db2f2010cd3bdee7cd1cd6d61d7297fd