L4::Vcon, L4::Scheduler: be more precise on IRQ
Both classes inherit from L4::Icu and L4::Irq_eoi. Document the limited
implementations of the respective virtual interrupt functionality.

The text is almost the same for both classes. I think this makes sense
because another class, L4::Semaphore, also implements a virtual device
interrupt but the behavior is different (there it is possible to change
the mode to level-triggered). Therefore it IMO does not make much sense
to create a dedicated chapter for virtual device interrupts.

See ticket #CD-478.

Change-Id: I7e5743fcbcb1886cfab305fbc946062aeca6668e