sigma0: Enforce page granular size for mappings
When requesting a memory mapping, either using map_anypage, map_mem
or map_iomem, then the reservation system will use the provided size.
However, since the kernel can only map at a minimum granularity of the
page size multiple reservations for different owners that are on the
same page may lead to handing out memory of a different owner due to
this implicit rounding.

Reject requests that do not have a size of at least the page size to
avoid giving the assumption that anything less is mapped.

Change-Id: I6c55c0b6ce877495e19b0e0d03f375d3653fbc4e