MIPS: fix setup of the 'GP' register
Changes for crt0 of Sigma0, Moe and Loader.

The GP register needs to be initialized at startup. The old code did not
correctly initialize the register for MIPS32. That was the reason that
MIPS code didn't work with recent compilers. The new code calls the
corresponding SETUP_GP instruction for MIPS32.

Furthermore, remove the '.cpload $25' instruction which is supposed to
load the GP register with some procedure-local context in offset to the
$25 / $t9 register. In PIC mode the $25 / $t9 register is supposed to
contain the address of the current function. This is done before __main
is called but for the crt0 context this is not required.

Change-Id: I69bd2cd5cd78da59df6b4a28f680b321ecec78a5