sigma0: check alignment before mapping
When sending an answer to a memory map request, the kernel will only
send memory starting from an address aligned to the order of the send
size. Check that the requested base address is correctly aligned,
otherwise reject the request as invalid.

Also, the manual masking of the bits in the raw fpage value can cause
confusion for readers of the code. Make the purpose explicit by using
the dedicated l4_fpage_memaddr function.

Change-Id: Ie7be64ae2b0ecdea3fef8b6226cad539e5df1de5