Only obtain caps for valid std::threads
If a std::thread is not joinable per definition this means that the
native handle is at its default initialized value, which is 0 for our
underlying type. Since we would assume the handle to be a tcb pointer
and dereference it that would lead to a null pointer dereference.

Return an invalid cap for such std::threads not backed by a native
thread.

Change-Id: I602e16df9579fa22d66a09f8255102d011be8b59