l4re_kernel: keep the 16-byte alignment on x86/32
The loader_thread() function expects to be called via the 'CALL'
instruction. We must not directly jump to that function via ex_regs
because that would destroy the 16-byte alignment.

This fixes problems with gcc versions >= 10 more aggressively using
SSE instructions. The stack alignment was also a problem with older
gcc versions but either no SSE instructions were executed or the
stack was alignment later "by accident".

Fixes ticket #CD-495.

Change-Id: I7a6816f4115183746049d944b8a33020774d300f