Counting_cap_alloc: Provide a thread-safe variant
Implement state with atomic operations to make the counting allocator
thread safe.

Care must be taken to not prematurely reallocate a capability slot that
just got unused. The thread safe variant will start with an initial
count of 2. The last reference will decrease the counter to 1, where
the slot is unmapped and only then is the slot marked as free.

Because the free hint update is racy, we still have to scan the
remaining slots if nothing was found after the hint. This should
practically never happen, but it is still important to ensure that
every possible slot is examined.

Co-authored-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>
Change-Id: I3140610b57e23d116c037e87dcc332b7b3593aa1