sigma0: add PIE support for arm(64) and x86_64
Because sigma0 does not use the uclibc startup, it must call
reloc_static_pie() itself. The startup code for amd64, arm and arm64
additionally needs to be adapted to not emit relocations which are
forbidden in the read-only text segment. It's not possible to provide
support for i386 because this architecture does not allow PC-relative
addressing without an existing stack.

Change-Id: I5738da0c7922bcb0eeee3ddf190a57b8c35d8183