sigma0: add PIE support for arm(64) and x86_64
Because sigma0 does not use the uclibc startup it must call
_reloc_static_pie itself. A dedicated stub is added for cases where the
function might not be present. Normally the stub would be provided by
crt1 but sigma0 does not link with that. The startup code for amd64,
arm and arm64 also needs to been adapted to not emit relocations which
are forbidden in the read-only text segment.

Change-Id: I5738da0c7922bcb0eeee3ddf190a57b8c35d8183