the kernel marks UTCB of killed threads as unused
The kernel might write a UTCB even after the corresponding thread was
deleted. The thread has to terminate itself and hence needs to be
scheduled. Depending on the priorities in the system, this can take some
time.

Define the unused UTCB field 'xfer timeout' where the kernel writes 0
when it is guaranteed that the UTCB is no longer written by the kernel.
Unfortunately we cannot use any of the 'user' fields because all of them
are already used for other purposes.

The far-term solution is probably that the kernel guarantees that a UTCB
is no longer touched when the corresponding system call returns.

See ticket #CD-518.

Change-Id: Ic7e8aa98d8f782b684445cc6f866df3bffae2c37