uclibc: add assertion to __l4_add_utcbs()
Free UTCBs for pthreads are maintained by a free list starting at
__pthread_first_free_handle. The function __l4_add_utcbs() adds virgin
UTCBs to that list if there are currently no free UTCBs available. Any
available UTCB would be forgotten because the current value of
__pthread_first_free_handle is not considered.

Add an assert to show this assumption, remove a superfluous type cast
and a superfluous line break.

Change-Id: I06e2995b52788e08e77d9f8fb2b3a7dcf8447841