provide function to align stack pointer for thread fns
On x86/AMD64 the stack is off by a machine word if a function is not
called using the 'call' instruction. This happens when a thread is
started using `ex_regs`.

Provide l4_align_stack_for_direct_fncall() which fixes the stack
alignment when entering a thread function using `ex_regs`.

Change-Id: Ib283c30c16411abf2993bb0257d8e85b200ec98d