l4sys: fix ARM64-specific vCPU host registers
The kernel uses the same architectural dependent host state for ARM and
ARM64. It contains only one tpidruro register, the one for the host. The
tpidruro register for the guest state is part of `l4_vcpu_regs_t` (vCPU
register state). This was changed in the corresponding kernel change
I81e10229dd2beb837ea96519e675e729411a5a54.

Changing L4_VCPU_STATE_VERSION is not required because we only adapt the
user-visible vCPU state to the actual kernel implementation. And ARM32
already uses the correct definition.

Change-Id: Ia2f17e8e7ceeab5321638db02e60eaada02156b1