Improve support for the RGB32 pixel format
The RGB32 pixel format has three 8-bit entries for red, green and blue
and the fourth byte is reserved (it is effectively just a padding to
speed up the memory accesses).

To avoid an intrusive change (our intention is not to support strange
pixel formats with reserved bits all over the place, our intention is
to support the RGB32 pixel format which has been the dominant one since
the late 1990s), we compute a padding pseudo component that represents
the tailing reserved bits in RGB32 and similar pixel formats when
needed.

The semantics of accessing the actual color components is unaffected.

Change-Id: I6dbecef3c2bcb6a195d71e814d8ea92159227f9c