Add dedicated stack PHDR on arm no-MMU
The stack is allocated as special section that is adjacent to the
data/bss section. This makes it possible to merge both logical sections
into a single MPU region in hardware. It also removes the burden of the
loader to allocate a separate stack frame from some dynamic memory
pool.

Change-Id: If7dc218004d41651fc5d986720536a6d4aa3d7fc