l4sys: 'enum l4_' types shall start with capital letter
The coding style requires a capital first letter of types and enum
types. There are still a few enum types starting with 'l4_' and having
the '_t' suffix (deprecated syntax).

However, the 'l4_msgtag_flags' type neither has a capital letter nor the
'_t' suffix and there is a function with the same name. Having the type
and the function the same name leads to wrong Doxygen links because the
name without '()' is not necessarily used for the type.

While at it, do also change l4_msgtag_protocol and L4_fpage_consts for
the same reason. These types are not used outside of this package,
except for links in the Doxygen documentation.

This change doesn't touch existing 'enum l4_xyz_t' types.

Change-Id: Iaf91456aff8e98dcdee6141eb91be318557478c2