l4sys, cxx: several Doxygen fixes
A couple of fixes for missing / wrong Doxygen documentation.

Worth to mention is fixing the usage of '/*@{*/' and '/*@}*/': With a
simple C-style comment, these commands are not detected by Doxygen. This
prevented (for example) proper documentation of the L4_TIMEOUT_xyz
constants. Use '/**@{*/' and '/**@}*/' instead. Analog for the C++
variant.

Change-Id: I82635bdda7472988953380a744213b85dac960f2