static_heap: allocate initial TLS from static heap
The static heap backend is all about removing dynamic allocations at
runtime. The TLS segment of the main thread is allocated very early in
the libc startup. Use the memory directly from our pre-allocated pool
to get rid of the last remaining dynamic allocation before entering
main().

The "--whole-archive" dance is required to get the weak
__libc_alloc_initial_tls symbol actually overridden by the static
library. Without it GNU ld will not re-consider the symbol even though
it was weak in the first place.

Change-Id: I26fd8651d88603686166d965719d9505665d0746