l4re_kernel: properly adjust stack for ex-regs function
Use the function l4_align_stack_for_direct_fncall() to properly align
the stack of a newly started function.

This fixes a stack alignment issue with the L4Re kernel compiled with
gcc-12 on AMD64.

For x86, it's not required to use a separate entry function as the stack
is now properly aligned.

Change-Id: Icdbffb6b8a4760eed275cacd4e2a081ac3aef6a4