l4sys: document when granting is not performed
If the resulting attributes of a flexpage as they would appear in the
receive window either don't have the L4_CAP_FPAGE_R bit (object caps)
or none of the L4_FPAGE_RWX bits (memory, IO ports), then the grant
operation is not performed: The flexpage is not removed from the sender
and not added to the receiver.

Document that in the L4_MAP_ITEM_GRANT flag.

See ticket #CD-689.

Change-Id: Ifb4bb37b3d0577d6d773279eb77a273e0ce2654a