arm64: install EL1 exception handler
If userland is running at EL1, setup the VBAR_EL1 register to an EL1
exception handler printing the ELR and the SPSR before going to sleep
forever. This is more convenient than just triggering a PF at 0.

Basically we set the exception handler from the startup code (crt1.S,
tramp-arm64.S). In the Assembler setup code we have to distinguish
between Moe and other tasks to check if the task is running at EL1:

- For Moe (EXCEPTION_EL1_DETECT_X1 defined), read the value of the X1
  register. It contains the Assembler instruction to be used to enter
  the kernel and is set by the kernel for new threads.
- Other tasks read the instruction from the KIP.

See ticket #CD-771.

Change-Id: Ibfd69b5a4816660d996f9d8e1e70c703b2ee1ed0