make virt_cap_alloc pointer visible for later adaption
Ldso has a small capability allocator used during application startup,
which is intended to be overwritten by the applications reference
counting capability allocator in libl4re-util.so.

However, the virt_cap_alloc symbol resided in private data section of
libld-l4.so and thus could not be changed when libl4re-util.so was
dynamically loaded.

Since the original symbol should remain internal to ldso and not be
subject to symbol resoultion this change defines an alias with a unique
name to be used in dynamic symbol resolution.

fixes ticket #CD-453

Change-Id: Id96be9abdd8102234e6ee381e1042b8d17308540