uclibc-minimal: force strict alignment
The minimal libc is also used for code where the MMU could be disabled,
for example in bootstrap. In that case, the generated code has to follow
natural memory alignment rules, for example a 16-byte access needs to be
16-byte aligned architectures.

The proper compile switch for this is -mstrict-align. This option is
only supported on ARM64. This becomes more urgent with gcc-12.

See ticket #CD-777.

Change-Id: Ia932697a17e9f9c9cfcfafb1d555aad5e7cf5aee