ldscripts: Avoid rebuilding because build info missing
Because crti_s.o and kip_addr.o are not part of an TARGET (ALLOBJS) or
INSTALL_TARGET their .cmd file is not added to DEPS in binary.inc, i.e.
their .cmd file is not included.
This causes them to be recompiled every whenever the ldscripts package
is built, even if none of their prerequisites have changed, so no recompilation
would be necessary:
 > [ldscripts] ... Compiling crti_s.o   because build info missing
 > [ldscripts] ... Compiling kip_addr.o   because build info missing

Change-Id: I4ad2fc80fa20391f56e749dc2fed4faf234212f8