libdl: Remove unused CRT0 make variable
The CRT0 make variable is unused since l4-bender was introduced in
change Ibcf58e25798201bc2802afdf37799703934c16ee. As part of this
transition, the CRT0 mechanism was replaced with with l4-bender flags
(-nocrt and -nocrt1) and options (Link_Start).

As it is not apparent why libdl.so should be built without crt, i.e. why
CRT0 was set in the first place, thus do not add -nocrt to LDFLAGS.

Change-Id: I0e92d3039bd4e1758ccda7a011f0a95a23956158