uclibc/ldso: clean up PERFORM_BOOTSTRAP_GOT ifdeferry
3 architectures currently define PERFORM_BOOTSTRAP_GOT: avr32, mips and
xtensa. A block of code that applies relative relocations in the
DL_START is disabled when PERFORM_BOOTSTRAP_GOT is defined, unless it's
avr32 or mips, effectively disabling it only for xtensa.
This may be simplified by removing the call to elf_machine_relative from
the xtensa PERFORM_BOOTSTRAP_GOT and always using common code.

Change-Id: I45b9ae1756f5d0b475e40f69dcff3fbd23e3349c
Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>