l4sys/l4re: Doxygen fixes
Certain functions copy the entire documentation of another function but
add certain parameters. As of Doxygen 1.9.3 it is required to use

  \copybrief other_function
  \param foo
  \param bar
  \copydetails other_function

instead of

  \param foo
  \param bar
  \copydoc other_function

to keep the parameters sorted. Also move the \pre statement after the
parameter documentation for the same reason, otherwise the expanded
preconditions will break the parameter list. Future versions of Doxygen
might behave better (current 1.9.6-git still behaves like 1.9.3) but
using separate commands for brief and for details is not wrong and in
fact this fix only applies to a few functions.

See ticket #CD-829.

Change-Id: I33d273be60a812b79a1c723cc6c5e7464252233b