l4util/amd64/bitops: use 64-bit instructions
Using the 64-bit instructions for bit operations allows to pass
immediate bit values up to 63 and to address bitfields > 4G.

This change also fixes compiler warnings triggered by certain
parameters: If an immediate value could be used because the bit number
is smaller 32, the Assembler warned about not specifying the operand
length because the old code did not specify it (for example 'btc' rather
than 'btcq'). The correct parameter length is forced by casting the bit
number parameter to a 64-bit type.

Change-Id: I3c3da45e6600363df64bb92aa69408baddab44ff