l4sys: add infinite loop function
Current compilers are allowed to optimize away `for (;;) {}` loops or
equivalent. We make use of such loops, e.g., for functions that are not
meant to return. These will then not do what one expects and lead to
subtle bugs.

This commit provides a new inline function `l4_infinite_loop()` that
compiles down to a `LABEL: jmp LABEL;` in pseudo-assembly, even when the
compiler is optimizing as aggressively as described.

Change-Id: If94597ee7d9fa2e13c1baf71c21a2bacc660f951