l4sys/arm: Clang fixes for syscalls
Use 'mvn #~(syscall)' rather than 'mov #syscall' because Clang doesn't
accept the latter. In both cases the same instruction is assembled.

This fixes the following build error with Clang:

  <inline asm>:48:3: error: invalid instruction, any one of the
                            following would fix this:
    mov   pc, #(-0x00000004-8)
    ^
  <inline asm>:48:14: note: invalid operand for instruction
    mov   pc, #(-0x00000004-8)
             ^
  <inline asm>:48:14: note: operand must be a register in range
                            [r0, r15]
    mov   pc, #(-0x00000004-8)
             ^
  <inline asm>:48:9: note: operand must be a register in range
                           [r0, r12] or r14
    mov   pc, #(-0x00000004-8)
          ^

Actually the error message is misleading because the actual problem is
the huge operand value, not the register. GNU as performs the conversion
automatically but Clang does not (tested with clang-14..clang-16).

Change-Id: I345d8a5038d6382ce5ebf6e8fd5d6865d2836bbf