l4re_kernel: cleanup init code; don't use init_priority 0
There is no need to use the init_priority 0 for initializing the global
data. A priority of 101 is sufficient. This prevents a warning because
init_priorities <= 100 are reserved for internal use only.

Also, there is no need to protect the initializer with an `_initialized`
variable because the initializer is implemented only once and hence will
be called only once.

This change extends Iee330400cf20e19be12b6a9aeb65e2d3f2d02e04.

Change-Id: I9bb61a37ea89bc554ef0c45a87c8d52acffc70a2