libpthread: repair initialization and remove obsolete CRT0
As of Ibcf58e25798201bc2802afdf37799703934c16ee the CRT0 variable is no
longer used. Remove it.

Repair the libpthread.so initialization by fixing the dependency to
so_init-$(ARCH).S. This corresponding object encodes a call to
__pthread_initialize_minimal() into the .init section. The code in the
.init section is executed after executing the _init function prolog in
crti.S and before the epilog of _init in crtn.S.

The old code did NOT add so_init-$(ARCH).s.o to libpthread.so at all.
Nevertheless the pthread library was initialized by an explicit call
from __uClibc_init(). But rather play safe by doing this from the _init
function as well.

The side effect of the missing dependency was that libpthread.so was
always rebuilt even if there were no changes in any dependent objects.

Change-Id: Ib4ebb324cc225b93f7ea7ab216a32f871667f1ee