l4sys: rewire arm64 syscall dispatching
On arm64 syscalls are made by jumping to the KIP. Now that the pointer
to the KIP is available in l4sys, we can rely on that pointer to jump
to the right address. As a first step, this is only used for dynamically
linked applications. For statically linked applications the linker
script still provides the __l4_sys_syscall symbol.

By utilizing the l4sys KIP pointer, the additional fixed mapping of the
KIP in ld.so can be removed.

Special care must be taken for the EL1 exception handler. The detection
to install it relies on the __l4_sys_syscall symbol which might be a
trampoline function now. This must now use l4_global_kip to find the
syscall instruction.

Change-Id: Ic499462d8bdcd72a0b835e934e6b9a632f82a057