moe: enable search for largest flexpage to unmap
Before this change, moe's search for the largest flexpage to unmap in a
contiguous dataspace was, most probably unwittingly, disabled by setting
the "hot spot" parameter of the address function to ~0. This change
removes the parameter to use the default of 0 which enables the search
and substantially decreases the number of unmap syscalls.

Fixes #CD-344.

Change-Id: I7f61744e3f1ecb0247f4656c2c1c9e9309071bde