moe: extend addressable CPU mask of scheduler proxy
The scheduler proxy factory method is extended to accept more than one
l4_umword_t as CPU mask. All supplied words are concatenated to form a
CPU mask that can address all available CPUs of the system.

Internally the scheduler proxy CPU masks are dynamically allocated with
the size of the underlying schedulers limit. But because the
L4::Scheduler API cannot address arbitrary sets of CPUs, it might not
be possible to fully forward the exact intersection between the
scheduler proxy mask and the scheduling parameters of a thread. This
should pose no problem, though. Moe utilizes the kernel scheduler
capability directly and the kernel will only pick the first CPU in the
CPU set anyway.

See ticket #CD-892.

Change-Id: Ib348c2a1544a289cc282af00209a1b9ccd565179