sigma0/moe: reintroduce syscall patching
Syscall patching is only done when the system is misconfigured. So only
when it is already clear it would not boot if not patched.

The patching always works for Sigma0, but for Moe it only works when the
W^X protection of moe is not enforced through bootstrap, which is the
default after I2010f37aab063cf4d021718561d11a5a6f714b56. Otherwise the
system would fail to boot due to trying to write executable memory. But
then it would fail to boot anyways due to the misconfiguration.

This change provides for a smother out-of-the box experience even in the
face of system configuration errors.

This reverts commit ad294c290819ca651b8441d20c1ec7f8946a28d1.

Change-Id: I581cfd54b913369e8161bc20e2fe9bf5b33ee4ac