Moe: try bigger mappings in dma_map
Try to dma map bigger regions at once by starting at the order of the
requested size. If no empty region is found in the virtual address
space, decrease the page order and try again.

This improves performance and decreases the memory amount required in
the kernel for the mappings.

Change-Id: If6b107e656d61bcc4a03d5993ea463c0eb795922