arm: always run threads in EL0 by default
On a !cpu_virt kernel, threads naturally run in EL0. We now keep the
same default for cpu_virt kernels too.

Dedicated support for this runtime model is needed in libpthread,
though. New threads wants to inherit the exception level which is an
explicit operation when an application runs in EL1.

Remove the facilities in sigma0 and moe that cope with the different
syscall conventions on arm. They are not required any more.

Change-Id: I2b949d873d541dbe16ebb124e53a258903030dcc