ned: correctly register L4::Irq handler
To receive an L4::Irq the object must be registered with
register_irq_obj(). Otherwise, when registered with register_obj(), the
signaling might seem to work, but only if the server is currently in an
(open) wait when the irq is triggered.

With respect to ned, this caused exit signals to be lost if the ned
main thread was currently executing while some child task terminates.

See ticket #CD-1005.

Change-Id: Iab36ed4bbe93e2b812a29f8ebc1819ec7e27239c