l4sys/ipc_iface: a 'long' error code must not overflow 'int'
There is an unlikely case that an RPC function returns an error code in
the label (type 'long') which is smaller than the smallest 'int' value.
In that case we must not simply cast the 'long' value to 'int' because
this might remove the negative sign from the error code and the caller
would interpret this return code as success. Analogously for return
codes larger than the largest 'int' value.

Introduce L4_EMSGERRRANGE for these cases.

This fixes gcc-13 warnings like the following one:

  ...include/l4/sys/cxx/types:240:14: warning:
    ‘tflags.L4Re::Rm::F::Flags::raw’ may be used uninitialized
    [-Wmaybe-uninitialized]
  240 |     { return DT(l.raw & r.raw); }
      |              ^~~~~~~~~~~~~~~~~
  In file included from ...l4/pkg/l4re-core/ldso/ldso/vfs.cc:127:
    ...include/l4/l4re_vfs/impl/vfs_impl.h:
      In member function ‘virtual int {anonymous}::Vfs::mremap(void*,
        size_t, size_t, int, void**)’:
    ...include/l4/l4re_vfs/impl/vfs_impl.h:782:23: note:
      ‘tflags.L4Re::Rm::F::Flags::raw’ was declared here
    782 |       L4Re::Rm::Flags tflags;
        |                       ^~~~~~

The following code illustrates the problem:

  #include <cstdio>
  #include <climits>
  int main()
  {
    long a = INT_MIN;
    printf("     a = %+ld / %#08lx\n", a, a);
    printf("(int)a = %+d / %#08x\n", (int)a, (int)a);
    long b = a - 1;
    printf("     b = %+ld / %#08lx\n", b, b);
    printf("(int)b = %+d / %#08x\n", (int)b, (int)b);
    return 0;
  }

On 64-bit host, the following output is generated -- note the last line
showing the problem.

       a = -2147483648 / 0xffffffff80000000
  (int)a = -2147483648 / 0x80000000
       b = -2147483649 / 0xffffffff7fffffff
  (int)b = +2147483647 / 0x7fffffff

Co-authored: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Co-authored: Jan Klötzke <jan.kloetzke@kernkonzept.com>
Change-Id: If207fe22d34da7949d5071c70541e8fe3b888481
Signed-off-by: Frank Mehnert <frank.mehnert@kernkonzept.com>