l4sys: clarify return tag and parameters of send-only IPCs
When performing an object invocation using a send-only IPC, we explain
that the returned message tag has mostly undefined content except of the
IPC error indicator (bit 15) which is set if the (send) IPC failed.

Do this consistently wherever we use a send-only IPC and also clarify
the cases where a send-only as well as a send-plus-receive operation is
possible.

Also don't write `\attention` because this might cause confusion. There
is nothing wrong using a send-only IPC.

And clarify the description of l4_ipc_error() and l4_error().

Change-Id: I8951b76c9f1df925b064c61266fb956e19d75995