ld.lld: l4_sys_syscall declaration compatible with LLVM ld
The syntax

  PROVIDE_HIDDEN(symbol = DEFINED(symbol) ? symbol : default_value);

does not work with ld.lld: The resulting `symbol` value is always 0. Use

  PROVIDE_HIDDEN(symbol = default_value);

instead. This works with GNU ld as well as with LLVM ld: The symbol is
only defined in the linker script if (1) it is referenced and (2) not
defined by any object included in the link -- this is exactly what's
required here.

Change-Id: I470a7fab86e3f155ba61de089a85d131edb8126c