fix resource leak on Dirinfo::create_ds_and_register()
When the destructor of Name_space is called, it iterates over the
entries registered with the namespace and calls delete on each of
those entries. The destructor of Moe::Entry only frees the
capability slots of the entries which it remembers to have
allocated itself. Such entries are marked with the F_allocated
flag.

Fixes ticket #CD-1100

Change-Id: I59115131a31601ac4e392658500435af8d4f38e6