uclibc: Fix -Warray-parameter warning for __sigsetjmp
Fix the [-Warray-parameter=] warning for __sigsetjmp generated by GCC 11 and
later GCC versions:
|
| warning: argument 1 of type 'struct __jmp_buf_tag *' declared as a pointer [-Warray-parameter=]
|  extern int __sigsetjmp (struct __jmp_buf_tag *__env, int __savemask) __THROWNL;
| ...
| note: previously declared as an array 'struct __jmp_buf_tag[1]'
|  extern int __sigsetjmp (struct __jmp_buf_tag __env[1], int __savemask)
|

Use the same fix as in glibc. The fix is to move the struct __jmp_buf_tag
definition to a separate bits/ header so it can be included in
pthread.h, to allow to use an array (as in setjmp.h) rather than a pointer
in the declaration.

Change-Id: I7b20f60b398aa418adce6cb82713781f9e302077
Signed-off-by: Pavel Kozlov <pavel.kozlov@synopsys.com>