uclibc: bits/byteswap-common.h: import recent headers from glibc
On systems where uClibc doesn't provide an arch specific byteswap.h,
we fall back on bits/byteswap-common.h.  However, there is a bug
in this header in the __bswap_constant_64(x) macro.  If, for example,
a double is passed, we get 'invalid operands to binary &' in which we
mismatch a 'double' and 'long long unsigned int'.  The newer glibc
headers fix this and so we import them.  Also, since the inclusion of
byteswap.h is not appropriate for assembly code, we guard a portion
of endian.h which uses byteswap.h from inclusion in any assembly.

This is needed, for example, for f2fs-tools 1.6.0 on 32-bit big
endian PowerPC.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>

KK-Note:

This partially reverts I508647d1c845f83bdffe678eab52f354438aed52 which
will be re-created in the commit on top.

Change-Id: I24f43b386336dc91d822e76d6dc76e60a86943f2